third_party/blink/web_tests/platform/generic/external/wpt/css/css-contain/container-queries/font-relative-units-dynamic-expected.txt

This is a testharness.js-based test.
PASS em units respond to changes
FAIL rem units respond to changes assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
PASS ex units respond to changes
PASS ch units respond to changes
Harness: the test ran to completion.
